+++ CHANGELOG.md
@@ -25,7 +25,7 @@
 
 **Changed**
 
-- Nothing yet!
+- Reduced the size of persisted fatal issue reports.
 
 **Fixed**